修改bug

+++ ephub-data-mysql-service/src/main/java/com/shangpin/ephub/data/mysql/hub/waitselect/mapper/HubWaitSelectMapper.java
@@ -16,6 +16,6 @@ import com.shangpin.ephub.data.mysql.hub.waitselect.po.HubWaitSelectResponse;
  */
 @Mapper
 public interface HubWaitSelectMapper {
-    int count(HubWaitSelectRequest hubWaitSelect);
+    Long count(HubWaitSelectRequest hubWaitSelect);
     List<HubWaitSelectResponse> selectList(HubWaitSelectRequestWithPage hubWaitSelectWithPage);
 }
